chatgpt_proxy/pyproject.toml
[project]
name = "chatgpt-proxy"
dynamic = ["version"]
description = "ChatGPT proxy server for UnrealScript / UE3 bots."
readme = "README.md"
requires-python = ">=3.13"
license-files = ["LICENSE"]
dependencies = [
    "asyncpg>=0.30.0",
    "click>=8.2.1",
    "loguru>=0.7.3",
    "openai>=1.88.0",
    "py-markdown-table>=1.3.0",
    "pyjwt>=2.10.1",
    "pypika>=0.48.9",
    "sanic[ext]>=25.3.0",
    "ujson>=5.10.0",
]

[build-system]
requires = ["hatchling", "uv-dynamic-versioning"]
build-backend = "hatchling.build"

[tool.hatch.version]
source = "uv-dynamic-versioning"

[tool.hatch.build.hooks.version]
path = "chatgpt_proxy/_version.py"

[dependency-groups]
dev = [
    "hatch>=1.14.1",
    "mypy>=1.16.1",
    "pytest>=8.4.1",
    "pytest-asyncio>=1.0.0",
    "respx>=0.22.0",
    "ruff>=0.12.0",
    "sanic-testing>=24.6.0",
    "uv-dynamic-versioning>=0.8.2",
]

[tool.hatch.build.targets.wheel]
exclude = [
    "chatgpt_proxy/tests"
]

# https://github.com/MagicStack/asyncpg/issues/387
# https://github.com/MagicStack/asyncpg/pull/577
[[tool.mypy.overrides]]
module = [
    "asyncpg.*",
]
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = [
    "pypika.*",
]
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = [
    "sanic_testing.*",
]
ignore_missing_imports = true
